requirements.txt
asgiref==3.8.1
Django==5.1.5
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
